Minor fixes
Thank you for putting this draft together. I've been wanting to throw together an implementation of Kyber for a while to learn a bit more about it, and this was the best resource I could find for it. I'll also drop some comments on areas that were a bit tougher to grok to see if there are any additional recommended changes I should fold in.